third_party/blink/web_tests/html/tabular_data/table_insertrow_default_argument-expected.txt

Tests that HTMLTableElement.insertRow()'s default argument is -1.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS table.__proto__ is HTMLTableElement.prototype
PASS rows.length is 4
PASS rows[0].innerHTML is "0"
PASS rows[1].innerHTML is "1"
PASS rows[2].innerHTML is "2"
PASS rows[3].innerHTML is "3"
PASS successfullyParsed is true

TEST COMPLETE
